+++ src/models/deviceSettings/sonosDeviceSettings.ts
@@ -1,0 +1,14 @@
+import { DeviceSettings } from './deviceSettings';
+
+export class SonosDeviceSettings extends DeviceSettings {
+  public maxPlayOnAllVolume: number = 80;
+
+  public fromPartialObject(data: Partial<SonosDeviceSettings>): void {
+    this.maxPlayOnAllVolume = data.maxPlayOnAllVolume ?? this.maxPlayOnAllVolume;
+    super.fromPartialObject(data);
+  }
+
+  protected toJSON(): string {
+    return JSON.stringify(this);
+  }
+}