+++ src/models/deviceSettings/tvSettings.ts
@@ -3,4 +3,13 @@ import { Daytime } from '../daytime';
 
 export class TvSettings extends DeviceSettings {
   public automaticTurnOff: Daytime | undefined;
+
+  public fromPartialObject(data: Partial<TvSettings>): void {
+    this.automaticTurnOff = data.automaticTurnOff ?? this.automaticTurnOff;
+    super.fromPartialObject(data);
+  }
+
+  protected toJSON(): string {
+    return JSON.stringify(this);
+  }
 }