+++ src/models/deviceSettings/wledSettings.ts
@@ -13,4 +13,24 @@ export class WledSettings extends DimmerSettings {
   public dayPreset?: number;
   public duskPreset?: number;
   public nightPreset?: number;
+
+  public fromPartialObject(data: Partial<WledSettings>): void {
+    this.dayOn = data.dayOn ?? this.dayOn;
+    this.dayBrightness = data.dayBrightness ?? this.dayBrightness;
+    this.dawnOn = data.dawnOn ?? this.dawnOn;
+    this.dawnBrightness = data.dawnBrightness ?? this.dawnBrightness;
+    this.duskOn = data.duskOn ?? this.duskOn;
+    this.duskBrightness = data.duskBrightness ?? this.duskBrightness;
+    this.nightOn = data.nightOn ?? this.nightOn;
+    this.nightBrightness = data.nightBrightness ?? this.nightBrightness;
+    this.dawnPreset = data.dawnPreset ?? this.dawnPreset;
+    this.dayPreset = data.dayPreset ?? this.dayPreset;
+    this.duskPreset = data.duskPreset ?? this.duskPreset;
+    this.nightPreset = data.nightPreset ?? this.nightPreset;
+    super.fromPartialObject(data);
+  }
+
+  protected toJSON(): string {
+    return JSON.stringify(this);
+  }
 }